add and document experimental options to make aggressive use of dpkg's trigger and configuration handling (Closes: #473461)
Add NoTriggers option to add --no-triggers to all dpkg calls,
NoConfiguration to prevent apt from trying to configure packages -
dpkg should handle this in the last ConfigurePending call.
This options are for now deactivated as they require more testing in
real world situations, but the plan is to enable them in the near
future if anything works well.